requirements.txt
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
fastapi==0.92.0
Flask==2.2.2
h11==0.14.0
hazelcast-python-client==5.1
httptools==0.5.0
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pydantic==1.10.5
python-dotenv==1.0.0
PyYAML==6.0
requests==2.28.2
sniffio==1.3.0
starlette==0.25.0
typing_extensions==4.5.0
urllib3==1.26.14
uuid==1.30
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
Werkzeug==2.2.2
zipp==3.13.0
